refactor: update server help info and reuse controller code
