+++ views/plugin/ogc/ui/ogcserverhelp.html
@@ -1,6 +1,6 @@
 <div id="time-help" class="flex-fill w-100 d-flex flex-column">
   <div class="modal-body container flex-fill">
-    Please provide a valid GetCapabilities URL for each supported service on the OGC server. The GetCapabilities URL can have many different formats, such as:
+    Please provide a valid GetCapabilities URL for each supported service on the OGC server.
     <br>
     <br>
     A GetCapabilities URL can have many different formats, such as:
@@ -15,11 +15,11 @@
     Instances of WMS, WFS, and WMTS in the URLs above are often interchangeable.
     <br>
     <br>
-    Supported WMS versions are: 1.1.1, 1.3.0
+    Supported WMS versions: 1.1.1, 1.3.0
     <br>
-    Supported WFS versions are: 1.1.0
+    Supported WFS versions: 1.1.0
     <br>
-    Supported WMTS versions are: 1.0.0
+    Supported WMTS versions: 1.0.0
   </div>
   <div class="modal-footer d-flex">
     <button class="btn btn-primary" ng-click="ctrl.close()">
